+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Only trigger diagnostic updates when changes have occurred.
+- Dispose of `Logger` correctly.
 
 ## [0.4.0] - 2021-02-21
 ### Added